refactor: authentication flow
