+++ app/(app)/(tabs)/picture.tsx
