+++ app/(app)/feed/detail/[entryId].tsx
